pyproject.toml
[tool.poetry]
name = "defense-quickstart-audio-transcription-translation"
version = "0.0.1"
description = "Transcribe war-related audio recordings and translate them into English with this example repository"
authors = ["Restack Team <service@restack.io>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
restack-ai = "^0.0.37"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-dotenv = "0.19"
streamlit = "^1.40.0"
requests = "^2.32.3"
restack-sdk-cloud = "^1.0.12"
groq = "^0.12.0"
openai = "^1.55.0"
torch = {version = "^2.5.1", platform = "darwin"}
transformers = {extras = ["torch"], version = "^4.46.3"}
assemblyai = "^0.35.1"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
services = "src.services:run_services"
app = "src.app:run_app"